Changed the korats lane field to an int instead of a float
Better space efficiency because all the lane positions are whole integers

+++ include/GameViewPlayer.h
@@ -37,11 +37,11 @@ public:
     sf::Texture gameSky;
     sf::Texture lockIcon;
 
-    float lane1 = 335;
-    float lane2 = 422;
-    float lane3 = 508;
-    float lane4 = 594;
-    float lane5 = 680;
+    int lane1 = 335;
+    int lane2 = 422;
+    int lane3 = 508;
+    int lane4 = 594;
+    int lane5 = 680;
