+++ src/Grunt.cpp
@@ -12,7 +12,7 @@
 /**
  * @brief Grunts spawn with 100 health.
  */
-Grunt::Grunt(float startLane) {
+Grunt::Grunt(int startLane) {
 	grunt.setSize(sf::Vector2f(25,50));
 	grunt.setOrigin(grunt.getSize().x / 2, grunt.getSize().y /2);
 	grunt.setPosition(1500,startLane);
@@ -47,12 +47,12 @@ void Grunt::moveGrunt(float timePassed)
     }
 }
 
-void Grunt::setLane(float givenLane)
+void Grunt::setLane(int givenLane)
 {
 	givenLane = lane;
 }
 
-float Grunt::getLane()
+int Grunt::getLane()
 {
 	return this->lane;
 }
